+++ pybaselines/utils.py
@@ -872,6 +872,31 @@ def optimize_window(*args, **kwargs):
     return estimate_window(*args, **kwargs)
 
 
+def _make_window(y, half_window):
+    """
+    Converts a half-window to full window for use with ndimage filters.
+
+    Parameters
+    ----------
+    y : numpy.ndarray, shape (N,) or shape (M, N)
+        The array of the measured data. Can be one or two dimensional.
+    half_window : int or numpy.ndarray([int, int]), optional
+        The half window size to use for the operations.
+
+    Returns
+    -------
+    window : numpy.ndarray, shape (2,) or list[int, ...]
+        The full window, with length matching the input y-dimensions for use
+        within SciPy's `ndimage` module.
+
+    """
+    window = 2 * half_window + 1
+    if isinstance(window, int):
+        window = [window] * y.ndim
+
+    return window
+
+
 def _avg_opening(y, half_window, opening=None):
     """
     Averages the dilation and erosion of a morphological opening on data.
@@ -897,12 +922,10 @@ def _avg_opening(y, half_window, opening=None):
     Raman Spectra of Artistic Pigments. Applied Spectroscopy, 2010, 64 595-600.
 
     """
-    window_size = 2 * half_window + 1
-    if isinstance(half_window, int):
-        window_size = [window_size] * y.ndim
+    window = _make_window(y, half_window)
     if opening is None:
-        opening = grey_opening(y, window_size)
-    return 0.5 * (grey_dilation(opening, window_size) + grey_erosion(opening, window_size))
+        opening = grey_opening(y, window)
+    return 0.5 * (grey_dilation(opening, window) + grey_erosion(opening, window))
 
 
 def _inverted_sort(sort_order):
